Add analysis to compute mappings between instructions and basic blocks.
This analysis builds a map from instructions to the basic block that
contains them.  It is accessed via get_instr_block().  Once built, it is kept
up-to-date by the IRContext, as long as instructions are removed via
KillInst.

I have not yet marked passes that preserve this analysis. I will do it
in a separate change.

Other changes:

- Add documentation about analysis values requirement to be powers of 2.
- Force a re-build of the def-use manager in tests.

+++ test/opt/ir_context_test.cpp
@@ -41,7 +41,7 @@ class DummyPassPreservesAll : public opt::Pass {
   Status Process(IRContext*) override { return status_to_return_; }
   Status status_to_return_;
   virtual Analysis GetPreservedAnalyses() override {
-    return Analysis(IRContext::kAnalysisEnd - 1);
+    return Analysis(static_cast<uint32_t>(IRContext::kAnalysisEnd) - 1);
   }
 };
